fix: fix code for CPU only execution
